fixing bugs on modifying gallery

+++ server/goadventures/src/main/java/io/softserve/goadventures/repositories/GalleryRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GalleryRepository extends CrudRepository<Gallery, Long> {
-    Gallery findById(long id);
+public interface GalleryRepository extends CrudRepository<Gallery, Integer> {
+    Gallery findById(int id);
     Gallery findByEventId(int eventId);
 }
